fix: update model loading for transformers v5.3+ API changes
Replace deprecated `torch_dtype` with `dtype` and add missing
`accelerate` dependency required by `device_map="auto"`.
Also mock `TranslateGemmaTranslator` in load model tests to prevent
unintended real model downloads during testing.

+++ src/translate_gemma_ui/translator.py
@@ -59,7 +59,7 @@ class TranslateGemmaTranslator:
         self._model = AutoModelForImageTextToText.from_pretrained(
             model_id,
             device_map="auto",
-            torch_dtype=dtype,
+            dtype=dtype,
             token=token,
         )
         self._max_tokens = getattr(self._model.config, "max_position_embeddings", 8192)
